refactor: restructure web package for shared module usage
Moved web-related classes and configurations from `undecided.erp.common.web` to `undecided.shared.web` to enhance reusability across modules. Updated all associated package references and resolved minor formatting inconsistencies.
