--- backend/src/main/java/undecided/erp/common/web/package-info.java
@@ -1,5 +1,0 @@
-/**
- * This package contains web-related components for the common module.
- * It includes configuration, exception handling, and logging for web requests and responses.
- */
-package undecided.erp.common.web;
