+++ backend/src/main/java/undecided/shared/web/exception/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * This package contains exception handling components for web requests. It includes exception
+ * resolvers, logging interceptors, and filters for handling and logging exceptions that occur
+ * during web request processing.
+ */
+package undecided.shared.web.exception;
