+++ backend/src/main/java/undecided/shared/web/logging/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * This package contains logging components for web requests and sessions. It includes interceptors
+ * and listeners for logging HTTP requests, responses, and session events to provide traceability
+ * and debugging information.
+ */
+package undecided.shared.web.logging;
