+++ backend/src/main/java/undecided/shared/web/package-info.java
@@ -1,0 +1,1 @@
+package undecided.shared.web;
